Add solution for day 21
Part 1 makes an assumption that the die won't roll back over to 1,
though it seems to work for the example and my input.

Part 2 I got stuck for a while because I was changing the player every
roll rather than every third roll. Memoizing the results really came in
handy here.